+++ public/admin/_app/immutable/nodes/19.BJLxav3M.js
@@ -1,0 +1,1 @@
+import"../chunks/h6qLnZTA.js";import"../chunks/idVUOa82.js";import"../chunks/DZZUfsFT.js";import{t as e}from"../chunks/MI3tMLZY.js";function t(t){e(t,{})}export{t as component};